Add ssl producer for workers as well
